Removed redundant files
