--- backend/controllers/login.js
@@ -1,18 +1,0 @@
-const User = require("../models/user");
-
-let login = (req, res, next) => {
-  let token = req.cookies.auth;
-  User.findByToken(token, (err, user) => {
-    if (err) throw err;
-    if (!user)
-      return res.json({
-        error: true,
-      });
-
-    req.token = token;
-    req.user = user;
-    next();
-  });
-};
-
-module.exports = { login };
