+++ backend/controllers/login/login.js
@@ -1,0 +1,57 @@
+const userSchema = require("../../models/user");
+const bcrypt = require("bcrypt");
+const jwt = require("jsonwebtoken");
+
+let login = (req, res, next) => {
+  let token = req.cookies.auth;
+  User.findByToken(token, (err, user) => {
+    if (err) throw err;
+    if (!user)
+      return res.json({
+        error: true,
+      });
+
+    req.token = token;
+    req.user = user;
+    next();
+  });
+};
+
+userSchema.methods.comparepassword = function (password, cb) {
+  bcrypt.compare(password, this.password, function (err, isMatch) {
+    if (err) return cb(next);
+    cb(null, isMatch);
+  });
+};
+
+userSchema.methods.generateToken = function (cb) {
+  var user = this;
+  var token = jwt.sign(user.email.toHexString(), confiq.SECRET);
+
+  user.token = token;
+  user.save(function (err, user) {
+    if (err) return cb(err);
+    cb(null, user);
+  });
+};
+
+userSchema.statics.findByToken = function (token, cb) {
+  var user = this;
+  jwt.verify(token, config.SECRET, function (err, decode) {
+    user.findOne({ _id: decode, token: token }, function (err, user) {
+      if (err) return cb(err);
+      cb(null, user);
+    });
+  });
+};
+
+userSchema.methods.deleteToken = function (token, cb) {
+  var user = this;
+
+  user.update({ $unset: { token: 1 } }, function (err, user) {
+    if (err) return cb(err);
+    cb(null, user);
+  });
+};
+
+module.exports = { login };
