+++ backend/models/user.js
@@ -1,6 +1,4 @@
 var mongoose = require("mongoose");
-const bcrypt = require("bcrypt");
-const jwt = require("jsonwebtoken");
 
 const userSchema = mongoose.Schema({
   firstName: {
@@ -49,42 +47,5 @@ const userSchema = mongoose.Schema({
     type: String,
   },
 });
-
-userSchema.methods.comparepassword = function (password, cb) {
-  bcrypt.compare(password, this.password, function (err, isMatch) {
-    if (err) return cb(next);
-    cb(null, isMatch);
-  });
-};
-
-userSchema.methods.generateToken = function (cb) {
-  var user = this;
-  var token = jwt.sign(user.email.toHexString(), confiq.SECRET);
-
-  user.token = token;
-  user.save(function (err, user) {
-    if (err) return cb(err);
-    cb(null, user);
-  });
-};
-
-userSchema.statics.findByToken = function (token, cb) {
-  var user = this;
-  jwt.verify(token, config.SECRET, function (err, decode) {
-    user.findOne({ _id: decode, token: token }, function (err, user) {
-      if (err) return cb(err);
-      cb(null, user);
-    });
-  });
-};
-
-userSchema.methods.deleteToken = function (token, cb) {
-  var user = this;
-
-  user.update({ $unset: { token: 1 } }, function (err, user) {
-    if (err) return cb(err);
-    cb(null, user);
-  });
-};
 
 module.exports = userSchema;
